Training Data from : The Temporal Focus of Campaign Communication (2020 JOP)

The replication data is sourced from "The Temporal Focus of Campaign Communication," authored by Stefan Müller, and was published in the Journal of Politics in 2022. This study primarily delves into the temporal emphasis of party manifestos. The dataset encompasses 5,858 annotated data entries from countries such as the United Kingdom, Ireland, Canada, Australia, New Zealand, and the United States. Its central objective is to compute the percentage of sentences or quasi-sentences referring to the past, present, or future. This differentiation is made based on two categories: "Prospective" and "Retrospective".

Usage

data("cc_muller")

Format

A data frame with 7 variables:

text

Content of the text.

sentence_id

Unique identifier for each sentence.

countryname

Country's name.

party

Associated political party of the text.

date

Date of the record.

class

Type or classification.

class_pro_retro

Classification as either 'Prospective' or 'Retrospective'.

Source

Data provided by the author and https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/7NP2XH

Examples

if (FALSE) {
data(cc_muller)
head(cc_muller)
}