.onAttach Function for the flaiR Package

The provided R code describes the .onAttach function for the flaiR package. This function is automatically invoked when the flaiR package is loaded. Its primary purpose is to set up and check the environment for the package and to display startup messages. .onAttach is triggered when the flaiR package gets loaded. It produces messages indicating the versions of Python and Flair in use and provides o ther details related to the package.

Usage

classify_stance2(text, classifier, threshold_score = NULL, threshold = FALSE)

Details

Specifying Python Path: The function starts by looking for the path of Python 3. If it doesn't find it, it stops the package load with an error message.

Checking Python Version: Next, the function checks whether the identified version of Python is Python 3. If it's not, it emits a warning.

Checking PyTorch Version: The function then checks if PyTorch is correctly installed and fetches its version information.

Checking Flair Version: It also checks if Flair is correctly installed and fetches its version.

Installation Status of Flair: If Flair isn't installed, the function attempts to install PyTorch and Flair automatically using pip commands. If the installation fails, it produces an error message.

Success Message: If all the checks pass, a message is displayed indicating that Flair can be successfully imported in R via flaiR.

Specifying Python Version for Use: Lastly, the function specifies which version of Python to use within R using the reticulate package.