flair: A package for adding flair to your displayed source code.

The flair package provides ...

flair returns a string with formatting wrappers (currently only html)

flair_rx(x, pattern, before = NULL, after = NULL, ...)

flair_quick(x, pattern, before = NULL, after = NULL, ...)

flair(x, pattern, ...)

flair_all(x, ...)

# S3 method for default
flair_all(x, ...)

# S3 method for with_flair
flair_all(x, ...)

flair_args(x, ...)

flair_funs(x, ...)

flair_input_vals(x, ...)

Arguments

x
A string object

pattern
A pattern to match. By default, this is a fixed pattern; use flair_rx for regular expressions.

before
String giving specific html tags to insert before matched text.

after
String giving specific html tags to insert after matched text.

...
Formatting style options, passed to txt_style

Value

A string with formatting wrappers.

Flair functions

The flair_* functions add decorative formatting to source code.

Decorate functions

The decorate functions execute source code and simultaneously prepare it to have flair added, to display when knitted.

Examples

if (FALSE) {
code_string <- "foo <- mean(1:10, na.rm = TRUE)"

code_string %>% flair("foo")

code_string %>% flair_args()

code_string %>% flair_funs(color = "red")
}