Highlight Entities with Specified Colors and Tag

This function highlights specified entities in a text string with specified background colors, font colors, and optional labels. Additionally, it allows setting a specific font type for highlighted text.

Usage

highlight_text(text, entities_mapping, font_family = "Arial")

Arguments

text

A character string containing the text to highlight.

entities_mapping

A named list of lists, with each sub-list containing:

words: A character vector of words to highlight.

background_color: A character string specifying the CSS color for the highlight background.

font_color: A character string specifying the CSS color for the highlighted text.

label: A character string specifying a label to append after each highlighted word.

label_color: A character string specifying the CSS color for the label text.

font_family

A character string specifying the CSS font family for the highlighted text and label. Default is "Arial".

Value

An HTML object containing the text with highlighted entities.

Examples

library(flaiR)
data("uk_immigration")
uk_immigration <- head(uk_immigration, 1)
tagger_ner <- load_tagger_ner("ner")
results <- get_entities(uk_immigration$text,
                        uk_immigration$speaker,
                        tagger_ner,
                        show.text_id = FALSE)
#> Error in data.table(doc_id = rep(doc_id, length(entities)), entity = vapply(entities,     function(e) e$text, character(1)), tag = vapply(entities,     function(e) e$tag, character(1))): could not find function "data.table"

highlighted_text <- highlight_text(uk_immigration$text, map_entities(results))
#> Error in eval(expr, envir, enclos): object 'results' not found
print(highlighted_text)
#> <div style="text-align: justify; font-family: Arial">Das <span style="background-color: pink; color: black; font-family: Arial">Bundesverfassungsgericht</span> <span style="color: pink; font-family: Arial">(ORG)</span> hat unterdessen die Aussetzung der Abschiebung von Asylsuchenden nach <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> im Rahmen des <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span>-Verteilungssystems, der Dublin-II-Verordnung, in acht Einzelfällen angeordnet. Das Gericht stützt sich dabei auf „ernst zu nehmende Quellen“, wonach eine ordnungsgemäße Registrierung als Asylsuchender in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> unmöglich sein könnte. Trotz der mittlerweile ergangenen acht einstweiligen Anordnungen des BundesverfassungsgeZu Protokoll gegebene Reden richts betreibt das <span style="background-color: pink; color: black; font-family: Arial">Bundesamt für Migration und Flüchtlinge</span> <span style="color: pink; font-family: Arial">(ORG)</span> weiterhin die Rückschiebung von Asylsuchenden nach <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span>. Dies wollen wir mit dem vorliegenden Antrag verhindern. Denn das Asylverfahren in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> ist weiterhin eine Frage des Zufallsprinzips. Nicht nur <span style="background-color: pink; color: black; font-family: Arial">Pro Asyl</span> <span style="color: pink; font-family: Arial">(ORG)</span> und <span style="background-color: pink; color: black; font-family: Arial">Human Rights Watch</span> <span style="color: pink; font-family: Arial">(ORG)</span>, sondern auch der <span style="background-color: pink; color: black; font-family: Arial">UNHCR</span> <span style="color: pink; font-family: Arial">(ORG)</span> berichten, dass das Asylverfahren in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> in vielerlei Hinsicht an erheblichen Mängeln leidet. Von einem fairen Verfahren, wie es nach dem internationalen Flüchtlingsrecht und den <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span>-Richtlinien über die Aufnahme von Flüchtlingen, die Durchführung des Asylverfahrens und die Kriterien für die Anerkennung als Flüchtling vorgesehen ist, kann man nicht sprechen. So kommen Inhaftierungen ohne Haftgrund vor, Dolmetscher bei der Befragung über die Fluchtgründe sind nicht garantiert, es gibt keine Unterbringung während des Asylverfahrens, der Zugang zur zentralen Asylbehörde in <span style="background-color: lightblue; color: black; font-family: Arial">Athen</span> <span style="color: blue; font-family: Arial">(LOC)</span> ist nur an einem einzigen Tag möglich. Dies alles räumt das <span style="background-color: pink; color: black; font-family: Arial">BMI</span> <span style="color: pink; font-family: Arial">(ORG)</span> auch in zahlreichen Stellungnahmen an den Petitionsausschuss des <span style="background-color: pink; color: black; font-family: Arial">Bundestages</span> <span style="color: pink; font-family: Arial">(ORG)</span> ein, will die Menschen aber dennoch weiter zurückschicken. Das ist aus grüner Sicht untragbar, denn <span style="background-color: lightblue; color: black; font-family: Arial">Deutschland</span> <span style="color: blue; font-family: Arial">(LOC)</span> trägt angesichts dieser dem Bundesinnenministerium schon länger bekannten Situation in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> gerade auch für rücküberstellte Personen aus <span style="background-color: lightblue; color: black; font-family: Arial">Deutschland</span> <span style="color: blue; font-family: Arial">(LOC)</span> eine Mitverantwortung. Aus unserer Sicht sollte <span style="background-color: lightblue; color: black; font-family: Arial">Deutschland</span> <span style="color: blue; font-family: Arial">(LOC)</span> die Asylverfahren hier in <span style="background-color: lightblue; color: black; font-family: Arial">Deutschland</span> <span style="color: blue; font-family: Arial">(LOC)</span> durchführen. Auch bei hohen Zugangszahlen von Asylantragstellern muss ein faires Verfahren unter Einhaltung der Mindeststandards aus der <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span>-Flüchtlingsaufnahme-Richtlinie, der <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span>-Asylverfahrens-Richtlinie und der EUQualifikations-Richtlinie erfolgen. Die südlichen Außengrenzländer der <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span> haben mit einer großen Zahl schutzsuchender Menschen zu tun: <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> vor allem mit Flüchtlingen aus dem <span style="background-color: lightblue; color: black; font-family: Arial">Irak</span> <span style="color: blue; font-family: Arial">(LOC)</span>, <span style="background-color: lightblue; color: black; font-family: Arial">Afghanistan</span> <span style="color: blue; font-family: Arial">(LOC)</span>, <span style="background-color: lightblue; color: black; font-family: Arial">Iran</span> <span style="color: blue; font-family: Arial">(LOC)</span>. Viele dieser Menschen haben schwerste Menschenrechtsverletzungen durchlitten und suchen nach einem sicheren Platz. <span style="background-color: lightblue; color: black; font-family: Arial">Deutschland</span> <span style="color: blue; font-family: Arial">(LOC)</span> sollte sich intensiv für eine Neuregelung der Verteilungsregelung innerhalb der <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span> einsetzen. Bis zu einer Neuregelung darf aber das Prinzip der „Verknappung von Zugangsmöglichkeiten zum Asylverfahren“ in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> nicht weiterpraktiziert werden. Denn es trifft Opfer von Menschenrechtsverletzungen. Dies ist nicht hinzunehmen. Bestätigt in dieser Haltung fühlen wir uns auch durch das <span style="background-color: pink; color: black; font-family: Arial">Bundesverfassungsgericht</span> <span style="color: pink; font-family: Arial">(ORG)</span>: Dieses hat erneut mit Beschluss vom 8. Dezember 2009 die Aussetzung der Abschiebung eines Asylsuchenden nach <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> im Rahmen des <span style="background-color: pink; color: black; font-family: Arial">EU</span> <span style="color: pink; font-family: Arial">(ORG)</span>-Verteilungssystems Nr. 343/2003 des Rates vom 18. Februar 2003, Dublin-II-Verordnung) angeordnet. Dafür war wie in dem der einstweiligen Anordnung vom 8. September 2009 – 2 BvQ 56/09 – zugrunde liegenden Fall ausschlaggebend, dass möglicherweise bereits mit der Abschiebung oder in ihrer Folge eintretende Rechtsbeeinträchtigungen nicht mehr verhindert oder rückgängig gemacht werden könnten. Zwar wird die Rückschiebung besonders schutzbedürftiger Flüchtlinge nach <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> in der Regel nicht vollzogen, die Argumentation des <span style="background-color: pink; color: black; font-family: Arial">Bundesamtes für Migration</span> <span style="color: pink; font-family: Arial">(ORG)</span> und Flüchtlinge, dass man von Asylbewerbern, die nicht besonders schutzbedürftig sind, erwarten könne, dass sie auch unter gegebenenfalls erschwerten Bedingungen das Asylverfahren in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> durchführten – Drucksache 16/14149 –, ist aber menschenrechtlich höchst bedenklich. Die Antragsteller der bisherigen positiven Eilverfahren vor dem <span style="background-color: pink; color: black; font-family: Arial">Bundesverfassungsgericht</span> <span style="color: pink; font-family: Arial">(ORG)</span> gehörten gerade nicht dem Kreis besonders schutzbedürftiger Personen an, bei denen die Bundesrepublik <span style="background-color: lightblue; color: black; font-family: Arial">Deutschland</span> <span style="color: blue; font-family: Arial">(LOC)</span> vom Selbsteintrittsrecht gemäß Art. 3 Abs. 2 der Dublin-IIVerordnung Gebrauch macht. Wenn aber das <span style="background-color: pink; color: black; font-family: Arial">Bundesverfassungsgericht</span> <span style="color: pink; font-family: Arial">(ORG)</span> die Verletzung elementarer Rechte in <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> für möglich hält und deswegen nach einer Abwägung die Rückführung unterbindet, darf sich die Bundesregierung dem nicht verschließen. Dennoch Rückführungen vorzunehmen, ist nicht nur eine Brüskierung des Bundesverfassungsgerichts, sondern heißt auch, die Menschenwürde der Asylsuchenden sehenden Auges zu gefährden. Daher fordern wir, Rückschiebungen nach <span style="background-color: lightblue; color: black; font-family: Arial">Griechenland</span> <span style="color: blue; font-family: Arial">(LOC)</span> im Rahmen des Dublin-II-Verfahrens sofort bis zur Hauptsacheentscheidung des Bundesverfassungsgerichts auszusetzen und die Prüfung der Asylanträge im Rahmen des Selbsteintritts im nationalen Asylverfahren durchzuführen.</div>