Load Flair POS Tagger

This function loads the POS (part-of-speech) tagger model for a specified language using the Flair library. If no language is specified, it defaults to 'pos-fast'.

Usage

load_tagger_pos(language = NULL)

Arguments

language

A character string indicating the desired language model. If NULL, the function will default to the 'pos-fast' model. Supported language models include:

"pos" - General POS tagging

"pos-fast" - Faster POS tagging

"upos" - Universal POS tagging

"upos-fast" - Faster Universal POS tagging

"pos-multi" - Multi-language POS tagging

"pos-multi-fast" - Faster Multi-language POS tagging

"ar-pos" - Arabic POS tagging

"de-pos" - German POS tagging

"de-pos-tweets" - German POS tagging for tweets

"da-pos" - Danish POS tagging

"ml-pos" - Malayalam POS tagging

"ml-upos" - Malayalam Universal POS tagging

"pt-pos-clinical" - Clinical Portuguese POS tagging

"pos-ukrainian" - Ukrainian POS tagging

Value

A Flair POS tagger model corresponding to the specified (or default) language.

Examples

if (FALSE) { # \dontrun{
tagger <- load_tagger_pos("pos-fast")
} # }